requirements_dev.txt
flake8-django==1.1.5
black==22.12.0
isort==5.12.0
drf-spectacular==0.28.0
factory_boy>=3.3.0
pytest>=7.0.0
pytest-django>=4.5.2
pytest-factoryboy>=2.0.0